+++ R/BinaryYScatterPlot.R
@@ -25,7 +25,8 @@
 #' @export
 BinaryYScatterPlot = function(frame, xvar, yvar,  title, ...,
                               se=FALSE, use_glm=TRUE) {
-  checkArgs(frame=frame,xvar=xvar,yvar=yvar,title=title,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::BinaryYScatterPlot")
+  checkArgs(frame=frame,xvar=xvar,yvar=yvar,title=title)
   frame[[yvar]] = as.numeric(frame[[yvar]])
   if(length(unique(frame[[yvar]])) != 2) stop(paste("outcome column", yvar, "not a binary variable"))