+++ R/ClevelandDotPlot.R
@@ -54,7 +54,8 @@ stemdotstats = function(ycol) {
 #' @export
 ClevelandDotPlot = function(frm, xvar, title, ...,
                             sort=-1, limit_n = NULL, stem=TRUE) {
-  checkArgs(frame=frm,xvar=xvar,yvar=xvar,title=title,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::ClevelandDotPlot")
+  checkArgs(frame=frm,xvar=xvar,yvar=xvar,title=title)
   if(!(is.null(limit_n) || isScalar(limit_n))) {
     stop("parameter limit_n must either be null or a numeric scalar")
   }