+++ R/DoubleDensity.R
@@ -24,7 +24,8 @@
 #'
 #' @export
 DoubleDensityPlot <- function(frame, xvar, truthVar,title,...) {
-  checkArgs(frame=frame,xvar=xvar,yvar=truthVar,title=title,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::DoubleDensityPlot")
+  checkArgs(frame=frame,xvar=xvar,yvar=truthVar,title=title)
   df <- data.frame(x=as.numeric(frame[[xvar]]),
                    y=as.character(frame[[truthVar]]),
                    stringsAsFactors=FALSE)