+++ R/GainCurve.R
@@ -104,13 +104,12 @@ GainCurvePlot = function(frame, xvar, truthVar, title,
                          ...,
                          compute_sig = TRUE,
                          large_count = 1000) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::GainCurvePlot")
   checkArgs(
     frame = frame,
     xvar = xvar,
     yvar = truthVar,
-    title = title,
-    ...
-  )
+    title = title)
   pctpop <- NULL # used as a symbol, declare not an unbound variable
   pct_outcome <-
     NULL # used as a symbol, declare not an unbound variable
@@ -322,13 +321,12 @@ GainCurvePlotC = function(frame, xvar, costVar, truthVar, title,
                           ...,
                           compute_sig = TRUE,
                           large_count = 1000) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::GainCurvePlotC")
   checkArgs(
     frame = frame,
     xvar = xvar,
     yvar = truthVar,
-    title = title,
-    ...
-  )
+    title = title)
   pctpop <- NULL # used as a symbol, declare not an unbound variable
   pct_outcome <-
     NULL # used as a symbol, declare not an unbound variable
@@ -528,13 +526,12 @@ GainCurvePlotWithNotation = function(frame,
                                      ...,
                                      compute_sig = TRUE,
                                      large_count = 1000) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::GainCurvePlotWithNotation")
    checkArgs(
     frame = frame,
     xvar = xvar,
     yvar = truthVar,
-    title = title,
-    ...
-  )
+    title = title)
   gainy = get_gainy(frame, xvar, truthVar, gainx)
   label = labelfun(gainx, gainy)
   gp = GainCurvePlot(frame, xvar, truthVar, title,