+++ R/LogLogPlot.R
@@ -32,6 +32,8 @@
 LogLogPlot <- function(frame, xvar, yvar, title,
                        ...,
                        use_coord_trans = FALSE) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::LogLogPlot")
+  checkArgs(frame=frame,xvar=xvar,yvar=yvar,title=title)
   XVAR <- NULL # don't look like an unbound variable
   YVAR <- NULL # don't look like an unbound variable
   linear_trend <- NULL # don't look like an unbound variable