+++ R/ScatterBoxPlot.R
@@ -24,7 +24,8 @@
 #' @export
 ScatterBoxPlot = function(frm, xvar, yvar, title, ...,
                           pt_alpha=0.3) {
-  checkArgs(frame=frm,xvar=xvar,yvar=yvar,title=title,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "WVPlots::ScatterBoxPlot")
+  checkArgs(frame=frm,xvar=xvar,yvar=yvar,title=title)
   if(!isDiscrete(frm[[xvar]])) {
     stop(paste(xvar, "should be discrete (factor, character, integer, or logical)"))
   }